use new assistant, remove re-creation of assistant
